Time: 89 ms (16.81%), Space: 30 MB (20.98%) - LeetHub

+++ 115-distinct-subsequences/115-distinct-subsequences.cpp
@@ -29,6 +29,6 @@ public:
                 
             }
         }
-       return dp[n][m];
+       return (int)dp[n][m];
     }
 };
